// src/types/card.d.ts
export type TierType = "normal" | "decent" | "amazing" | "elite" | "fantasy" | "goat";
export type PositionType =
  | "GK"
  | "CB"
  | "LB"
  | "RB"
  | "LWB"
  | "RWB"
  | "CDM"
  | "CM"
  | "CAM"
  | "LM"
  | "RM"
  | "LW"
  | "RW"
  | "LF"
  | "RF"
  | "CF"
  | "ST";

export type MomType = "AM" | "DE" | "EL" | "FA" | "GO" | "NO";

export type GradeType = "AM" | "DE" | "EL" | "FA" | "GO" | "NO" | "plate";

export type CardMode = "ingame" | "simulation";